+++ packages/expo-kline-chart/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @expo-kline-chart/core
+# expo-kline-chart
 
 ## 0.1.4
 
 ### Patch Changes
 
-- Add keywords to package.json for better npm discoverability
+-   Add keywords to package.json for better npm discoverability